# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Articles API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/V2/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category information by id",
                "description": "Get category information by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}